+++ demo/src/demo/windmill.cpp
@@ -48,7 +48,7 @@ public:
         {
             if (obstacles)
             {
-                RigidBody* c = world->CreateRegularPolygon(0.18f, Rand(3, 8));
+                RigidBody* c = world->CreateRegularPolygon(0.18f, (int32)Rand(3, 8));
                 c->SetPosition(RandVec2(Vec2{ -2.0f, 6.0f }, Vec2{ 2.0f, 6.0f }));
             }
